refactor: 폴더 구조 변경 및 폴더와 컴포넌트명이 일치하도록 변경
- index.tsx로 작성 통일
- style 파일도 폴더별로 구분
- static 파일을 ts 파일로 변경
