+++ src/App.tsx
@@ -13,7 +13,7 @@ import { Routes, Route } from 'react-router-dom';
 import ChangePassword from 'pages/Auth/FindIDPassword/page/ChangePassword';
 import ProtectedRoute from 'components/common/ProtectedRoute';
 import { Suspense } from 'react';
-import SearchDetailList from 'pages/SearchDetails';
+import SearchDetails from 'pages/SearchDetails';
 
 function App(): JSX.Element {
   return (
@@ -32,7 +32,7 @@ function App(): JSX.Element {
         </Route>
         <Route path="/post" element={<Post />} />
         <Route path="/search" element={<Search />} />
-        <Route path="/search/:searchQuery" element={<SearchDetailList />} />
+        <Route path="/search/:searchQuery" element={<SearchDetails />} />
       </Routes>
       <Toast />
     </Suspense>
