+++ src/pages/Search/components/NavigationBar/NavigationBar.module.scss
@@ -1,0 +1,42 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+.search {
+  width: 100%;
+  height: 100vh;
+  font-size: 32px;
+
+  @include media.media-breakpoint-up(mobile) {
+    max-width: 767px;
+    margin: 0 auto;
+  }
+}
+
+.search-nav {
+  display: flex;
+  padding-left: 30px;
+  padding-top: 72px;
+  align-items: center;
+
+  &__text {
+    font-size: 18px;
+    font-weight: 500;
+    height: 22px;
+    padding-left: 16px;
+  }
+
+  &__button {
+    &--previous {
+      display: flex;
+      flex-direction: column;
+      cursor: pointer;
+      background: none;
+      border: none;
+    }
+  }
+}
+
+.search-wrapper {
+  &--search {
+    padding-bottom: 24px;
+  }
+}
