+++ src/pages/Search/components/RelatedSearches/RelatedSearches.module.scss
@@ -1,0 +1,51 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+.search {
+  width: 100%;
+  height: 100vh;
+  font-size: 32px;
+
+  @include media.media-breakpoint-up(mobile) {
+    max-width: 767px;
+    margin: 0 auto;
+  }
+}
+
+// 검색창 결과 리스트
+.search-related-item {
+  height: calc(100vh - 180px);
+  overflow: scroll;
+
+  .link {
+    color: "black";
+    text-decoration: "none";
+  }
+
+  &__title {
+    display: flex;
+    font-size: 14px;
+    height: 34px;
+    padding-left: 52px;
+    align-items: center;
+    cursor: pointer;
+
+    &:hover {
+      background: #eeeeee;
+    }
+  }
+
+  &__text--not-found {
+    word-break: keep-all;
+    width: 275px;
+    height: 40px;
+    font-size: 16px;
+    line-height: 20px;
+    text-align: center;
+    margin: 0 auto;
+    margin-top: 164px;
+  }
+
+  &--hidden {
+    display: none;
+  }
+}
