+++ src/pages/Search/components/RelatedSearches/index.tsx
@@ -1,8 +1,8 @@
 import cn from 'utils/ts/classNames';
-import styles from 'pages/Search/Search.module.scss';
+import styles from 'pages/Search/components/RelatedSearches/RelatedSearches.module.scss';
 import MODE from 'pages/Search/static/mode';
 import suggestion from 'pages/Search/static/suggestion';
-import RelatedItem from './RelatedItem';
+import RelatedItem from './components/RelatedItem';
 
 interface Props {
   mode: string,
@@ -13,8 +13,8 @@ function RelatedSearches({ mode, text }: Props) {
   return (
     <div className={styles.search}>
       <ul className={cn({
-        [styles['search-query-list']]: true,
-        [styles['search-query-list--hidden']]: mode === MODE.trending,
+        [styles['search-related-item']]: true,
+        [styles['search-related-item--hidden']]: mode === MODE.trending,
       })}
       >
         {text === '' ? null : suggestion.filter((item) => item.title.includes(text))
