+++ src/pages/Search/components/SearchBar/RollingBanner.tsx
@@ -1,4 +1,4 @@
-import styles from 'pages/Search/Search.module.scss';
+import styles from 'pages/Search/components/SearchBar/SearchBar.module.scss';
 
 interface Props {
   trendings: string[]
@@ -6,12 +6,12 @@ interface Props {
 
 export default function RollingBanner({ trendings }: Props) {
   return (
-    <div className={styles['search-banner']}>
+    <div className={styles['search-rolling-banner']}>
       <strong>TRENDING</strong>
-      <ul className={styles['search-banner__tag-list']}>
-        <li className={styles['search-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
-        <li className={styles['search-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
-        <li className={styles['search-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
+      <ul className={styles['search-rolling-banner__tag-list']}>
+        <li className={styles['search-rolling-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
+        <li className={styles['search-rolling-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
+        <li className={styles['search-rolling-banner__tag']}>{trendings.map((tag) => `#${tag}`).join(' ')}</li>
       </ul>
     </div>
   );
