+++ src/pages/Search/components/SearchBar/SearchBar.module.scss
@@ -1,0 +1,97 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+.search {
+  width: 100%;
+  height: 100vh;
+  font-size: 32px;
+
+  @include media.media-breakpoint-up(mobile) {
+    max-width: 767px;
+    margin: 0 auto;
+  }
+
+  &__recommend {
+    word-break: keep-all;
+    height: 80px;
+    padding-top: 64px;
+    padding-right: 60px;
+    padding-left: 30px;
+    font-size: 32px;
+    font-weight: 700;
+    line-height: 40px;
+    white-space: pre-wrap;
+    color: #000000;
+  }
+}
+
+.search-bar {
+  display: flex;
+  border-radius: 15px;
+  border: none;
+  height: 30px;
+  align-items: center;
+  margin-top: 32px;
+  margin-right: 16px;
+  margin-left: 16px;
+  background: #eeeeee;
+  padding-left: 15px;
+
+  &__input {
+    width: 100%;
+    height: 15px;
+    border: none;
+    background: none;
+    font-size: 14px;
+    line-height: 15px;
+
+    &::placeholder {
+      color: #c4c4c4;
+      font-size: 12px;
+    }
+  }
+
+  &__icon {
+    width: 18px;
+    height: 18px;
+    padding-left: 15px;
+    padding-right: 14px;
+    cursor: pointer;
+  }
+}
+
+.search-rolling-banner {
+  display: flex;
+  overflow: hidden;
+  margin-top: 8px;
+  margin-right: 27px;
+  margin-left: 27px;
+  height: 15px;
+  font-size: 12px;
+
+  strong {
+    margin-right: 16px;
+  }
+
+  &__tag-list {
+    display: flex;
+    flex: 0 0 auto;
+    white-space: nowrap;
+    overflow: hidden;
+    list-style: none;
+
+    @keyframes text-loop {
+      0% {
+        transform: translate3d(0, 0, 0);
+      }
+
+      100% {
+        transform: translate3d(-100%, 0, 0);
+      }
+    }
+  }
+
+  &__tag {
+    animation: text-loop 10s linear infinite;
+    padding-right: 10px;
+  }
+}
