+++ src/pages/Search/static/mode.ts
