+++ src/pages/Search/static/recommend.ts
