+++ src/pages/Search/static/suggestion.ts
