+++ src/pages/Search/static/trend.ts
