+++ src/pages/SearchDetails/components/LoadingView.tsx
@@ -1,0 +1,7 @@
+import styles from 'pages/SearchDetails/SearchDetails.module.scss';
+
+export default function LoadingView() {
+  return (
+    <div className={styles['search__loading-view']}>현재 위치에서 찾아보는 중이에요..</div>
+  );
+}
